Move new regex validation helpers to a new class
